Gbe 296: Fix Update profile flow (#312)
* basically done

* update tests for act flow

And remove 11 redundant tests and combine the files.  The number of tests is getting crazy.

* fix up class tests

Removed over 10 other tests, and moved many to a central make_bid_test - in the hopes of removing more later.

Added in the needed tests for the make class changes with first/last name

* clean up test errors

* put in tests for new work

* prototype of test acceleration

Took the tests I was working with and the 5 test sets with the most cases, and move them to setTestUp pattern to set up the data less frequently.

* coverage

* pepdiff
