+++ templates/form_table.tmpl
@@ -42,7 +42,7 @@
     <div class="{% if two_col %}col-md-7{% else %}col-md-8{% endif %} form_field">
       {{ field }}
       {% if field.errors %}
-        <font class="gbe-form-error">{{ field.errors }}</font>
+        <br><font class="gbe-form-error">{{ field.errors }}</font>
       {% endif %}
     </div>
   </div>
