+++ tests/gbe/test_make_summer_act.py
@@ -40,6 +40,8 @@ class TestSummerAct(TestCase):
         form_dict = {'theact-shows_preferences': [4],
                      'theact-b_title': 'An act',
                      'theact-phone': '111-222-3333',
+                     'theact-first_name': 'Jane',
+                     'theact-last_name': 'Smith',
                      'theact-track_title': 'a track',
                      'theact-track_artist': 'an artist',
                      'theact-b_description': 'a description',
